requirements.txt
PyYAML==6.0
ujson==5.7.0
ultralytics==8.0.90
